Applied our formatting.

+++ src/theory/transformations/PushQuantifications.cpp
@@ -6,7 +6,7 @@
  * Written by Broes De Cat, Stef De Pooter, Johan Wittocx
  * and Bart Bogaerts, K.U.Leuven, Departement Computerwetenschappen,
  * Celestijnenlaan 200A, B-3001 Leuven, Belgium
-****************************************************************/
+ ****************************************************************/
 
 #include "IncludeComponents.hpp"
 #include "PushQuantifications.hpp"
@@ -16,9 +16,9 @@
 using namespace std;
 
 template<class VarList>
-bool sharesVars(const VarList& listone, const VarList& listtwo){
-	for(auto i=listone.cbegin(); i!=listone.cend(); ++i) {
-		if(listtwo.find(*i)!=listtwo.cend()){
+bool sharesVars(const VarList& listone, const VarList& listtwo) {
+	for (auto i = listone.cbegin(); i != listone.cend(); ++i) {
+		if (listtwo.find(*i) != listtwo.cend()) {
 			return true;
 		}
 	}
@@ -26,15 +26,15 @@ bool sharesVars(const VarList& listone, const VarList& listtwo){
 }
 
 template<typename FormulaList, typename VarList, typename VarSet>
-void splitOverSameQuant(const VarSet& quantset, const FormulaList& formulasToSplit, vector<FormulaList>& splitformulas, VarList& splitvariables){
+void splitOverSameQuant(const VarSet& quantset, const FormulaList& formulasToSplit, vector<FormulaList>& splitformulas, VarList& splitvariables) {
 	// TODO analyse van de beste volgorde => ordenen volgens stijgend aantal formules dat die variabele bevat.
-	for(auto i=quantset.cbegin(); i!=quantset.cend(); ++i) {
+	for (auto i = quantset.cbegin(); i != quantset.cend(); ++i) {
 		splitvariables.push_back(*i);
 	}
 	splitformulas.resize(splitvariables.size());
-	for(auto i=formulasToSplit.cbegin(); i<formulasToSplit.cend(); ++i) {
-		for(int j=splitvariables.size()-1; j>-1; --j){
-			if((*i)->freeVars().find(splitvariables[j])!=(*i)->freeVars().cend()){
+	for (auto i = formulasToSplit.cbegin(); i < formulasToSplit.cend(); ++i) {
+		for (size_t j = splitvariables.size() - 1; j > -1; --j) {
+			if ((*i)->freeVars().find(splitvariables[j]) != (*i)->freeVars().cend()) {
 				splitformulas[j].push_back(*i);
 				break;
 			}
@@ -49,45 +49,46 @@ Formula* PushQuantifications::visit(QuantForm* qf) {
 	quantforms.push_back(qf);
 	auto currentquant = qf;
 
-	while(true){
+	while (true) {
 		auto qsubf = dynamic_cast<QuantForm*>(currentquant->subformula());
-		if(qsubf!=NULL){
+		if (qsubf != NULL) {
 			quantforms.push_back(qsubf);
 			currentquant = qsubf;
-		}else{
+		} else {
 			auto bsubf = dynamic_cast<BoolForm*>(currentquant->subformula());
-			if(bsubf==NULL){
+			if (bsubf == NULL) {
 				return qf;
 			}
-			vector<vector<Formula*> > formulalevels(quantforms.size()+1); // Level at which they should be added (index 0 is before the first quantifier, ...)
+			vector<vector<Formula*> > formulalevels(quantforms.size() + 1); // Level at which they should be added (index 0 is before the first quantifier, ...)
 			conj = bsubf->conj();
 			Assert(bsubf->sign()==SIGN::POS);
-			for(auto i=bsubf->subformulas().cbegin(); i<bsubf->subformulas().cend(); ++i) {
+			for (auto i = bsubf->subformulas().cbegin(); i < bsubf->subformulas().cend(); ++i) {
 				auto subvars = (*i)->freeVars();
 				int index = quantforms.size();
-				while(index>0 && not sharesVars(subvars, quantforms[index-1]->quantVars())){
+				while (index > 0 && not sharesVars(subvars, quantforms[index - 1]->quantVars())) {
 					index--;
 				}
 				formulalevels[index].push_back(*i);
 			}
 			// Construct new formula:
 			Formula* prevform = NULL;
-			for(int i=quantforms.size()-1; i>-1; --i){
+			for (size_t i = quantforms.size() - 1; i > -1; --i) {
 				vector<vector<Formula*> > splitformulalevels;
 				vector<Variable*> splitvariables;
-				splitOverSameQuant(quantforms[i]->quantVars(), formulalevels[i+1], splitformulalevels, splitvariables);
+				splitOverSameQuant(quantforms[i]->quantVars(), formulalevels[i + 1], splitformulalevels, splitvariables);
 				Assert(splitvariables.size()>0);
-				for(int j=splitvariables.size()-1; j>-1; --j){
+				for (size_t j = splitvariables.size() - 1; j > -1; --j) {
 					auto subformulas = splitformulalevels[j];
-					if(prevform!=NULL){
+					if (prevform != NULL) {
 						subformulas.push_back(prevform);
-					}else{
+					} else {
 						Assert(subformulas.size()>0);
 					}
-					if(subformulas.size()==1){
-						prevform = new QuantForm(SIGN::POS, quantforms[i]->quant(), {splitvariables[j]}, subformulas.back(), quantforms[i]->pi());
-					}else{
-						prevform = new QuantForm(SIGN::POS, quantforms[i]->quant(), {splitvariables[j]}, new BoolForm(SIGN::POS, conj, subformulas, bsubf->pi()), quantforms[i]->pi());
+					if (subformulas.size() == 1) {
+						prevform = new QuantForm(SIGN::POS, quantforms[i]->quant(), { splitvariables[j] }, subformulas.back(), quantforms[i]->pi());
+					} else {
+						prevform = new QuantForm(SIGN::POS, quantforms[i]->quant(), { splitvariables[j] },
+								new BoolForm(SIGN::POS, conj, subformulas, bsubf->pi()), quantforms[i]->pi());
 					}
 				}
 			}
@@ -96,9 +97,9 @@ Formula* PushQuantifications::visit(QuantForm* qf) {
 			subformulas.push_back(prevform);
 
 			Formula* result = NULL;
-			if(subformulas.size()==1){
+			if (subformulas.size() == 1) {
 				result = subformulas.back();
-			}else{
+			} else {
 				result = new BoolForm(SIGN::POS, conj, subformulas, bsubf->pi());
 			}
